initial init

+++ backend/database/migrations/2024_08_24_053826_create_pembayarans_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('pembayarans', function (Blueprint $table) {
             $table->id();
             $table->foreignId('penjualan_id')->constrained('penjualans');
-            $table->decimal('amount');
+            $table->bigInteger('amount');
             $table->date('payment_date');
             $table->string('status');
             $table->timestamps();
